pyproject.toml
[project]
name = "aws-cdk-lightsail-openclaw"
version = "0.1.0"
description = "AWS CDK (Python + uv) project to provision OpenClaw on AWS Lightsail"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aws-cdk-lib>=2.190.0",
  "constructs>=10.3.0",
  "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
  "pytest>=8.4.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["tests"]
